+++ internal/service/mfa/resource_mfa_policy.go
@@ -4,12 +4,15 @@ import (
 	"context"
 	"fmt"
 	"net/http"
+	"time"
 
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/patrickcping/pingone-go-sdk-v2/management"
 	"github.com/patrickcping/pingone-go-sdk-v2/mfa"
+	"github.com/patrickcping/pingone-go-sdk-v2/pingone/model"
 	client "github.com/pingidentity/terraform-provider-pingone/internal/client"
 	"github.com/pingidentity/terraform-provider-pingone/internal/framework"
 	"github.com/pingidentity/terraform-provider-pingone/internal/sdk"
@@ -506,7 +509,8 @@ func resourceMFAPolicyDelete(ctx context.Context, d *schema.ResourceData, meta i
 
 	var diags diag.Diagnostics
 
-	_, diags = sdk.ParseResponse(
+	timeout := 30 * time.Second
+	_, diags = sdk.ParseResponseWithCustomTimeout(
 		ctx,
 
 		func() (any, *http.Response, error) {
@@ -514,8 +518,22 @@ func resourceMFAPolicyDelete(ctx context.Context, d *schema.ResourceData, meta i
 			return framework.CheckEnvironmentExistsOnPermissionsError(ctx, p1Client.API.ManagementAPIClient, d.Get("environment_id").(string), nil, fR, fErr)
 		},
 		"DeleteDeviceAuthenticationPolicy",
-		sdk.DefaultCustomError,
-		nil,
+		sdk.CustomErrorResourceNotFoundWarning,
+		func(ctx context.Context, r *http.Response, p1Error *model.P1Error) bool {
+
+			if p1Error != nil && r.StatusCode == http.StatusConflict {
+				if details, ok := p1Error.GetDetailsOk(); ok && details != nil && len(details) > 0 {
+					if target, ok := details[0].GetTargetOk(); ok && details[0].GetCode() == "INVALID_VALUE" && *target == "signOnPolicies" {
+						tflog.Warn(ctx, "Conflict detected when deleting MFA policy.  This is likely due to the policy being referenced by a sign-on policy.  Retrying..")
+
+						return true
+					}
+				}
+			}
+
+			return false
+		},
+		timeout,
 	)
 	if diags.HasError() {
 		return diags